+++ src/main/java/org/apache/mahout/clustering/canopy/CanopyDriver.java
@@ -29,8 +29,8 @@ public class CanopyDriver {
     String input = args[0];
     String output = args[1];
     String measureClassName = args[2];
-    float t1 = new Float(args[3]);
-    float t2 = new Float(args[4]);
+    double t1 = new Double(args[3]);
+    double t2 = new Double(args[4]);
     runJob(input, output, measureClassName, t1, t2);
   }
 
@@ -44,7 +44,7 @@ public class CanopyDriver {
    * @param t2               the T2 distance threshold
    */
   public static void runJob(String input, String output,
-                            String measureClassName, float t1, float t2) {
+                            String measureClassName, double t1, double t2) {
     JobClient client = new JobClient();
     JobConf conf = new JobConf(
             org.apache.mahout.clustering.canopy.CanopyDriver.class);
